$ Create component quick bookmark button to avoid duplicate code
