+++ src/renderer/components/ft-playlist-quick-bookmark-button/ft-playlist-quick-bookmark-button.vue
@@ -1,0 +1,17 @@
+<template>
+  <ft-icon-button
+    :title="quickBookmarkIconText"
+    :icon="isInQuickBookmarkPlaylist ? ['fas', 'check'] : ['fas', 'bookmark']"
+    class="quickBookmarkVideoIcon"
+    :class="{
+      bookmarked: isInQuickBookmarkPlaylist,
+      alwaysVisible: alwaysVisible,
+    }"
+    :theme="quickBookmarkIconTheme"
+    :padding="padding"
+    :size="size"
+    @click="toggleQuickBookmarked"
+  />
+</template>
+
+<script src="./ft-playlist-quick-bookmark-button.js" />
