+++ src/renderer/components/watch-video-info/watch-video-info.js
@@ -4,7 +4,8 @@ import FtCard from '../ft-card/ft-card.vue'
 import FtIconButton from '../ft-icon-button/ft-icon-button.vue'
 import FtShareButton from '../ft-share-button/ft-share-button.vue'
 import FtSubscribeButton from '../ft-subscribe-button/ft-subscribe-button.vue'
-import { formatNumber, openExternalLink, showToast } from '../../helpers/utils'
+import FtPlaylistQuickBookmarkButton from '../ft-playlist-quick-bookmark-button/ft-playlist-quick-bookmark-button.vue'
+import { formatNumber, openExternalLink } from '../../helpers/utils'
 
 export default defineComponent({
   name: 'WatchVideoInfo',
@@ -12,7 +13,8 @@ export default defineComponent({
     'ft-card': FtCard,
     'ft-icon-button': FtIconButton,
     'ft-share-button': FtShareButton,
-    'ft-subscribe-button': FtSubscribeButton
+    'ft-subscribe-button': FtSubscribeButton,
+    'ft-playlist-quick-bookmark-button': FtPlaylistQuickBookmarkButton,
   },
   props: {
     id: {
@@ -105,11 +107,6 @@ export default defineComponent({
     }
   },
   emits: ['change-format', 'pause-player', 'set-info-area-sticky', 'scroll-to-info-area'],
-  data() {
-    return {
-      selectQuickBookmarkTargetPromptCloseCallback: null,
-    }
-  },
   computed: {
     hideSharingActions: function() {
       return this.$store.getters.getHideSharingActions
@@ -221,48 +218,6 @@ export default defineComponent({
     defaultPlayback: function () {
       return this.$store.getters.getDefaultPlayback
     },
-
-    quickBookmarkPlaylistId() {
-      return this.$store.getters.getQuickBookmarkTargetPlaylistId
-    },
-    quickBookmarkPlaylist() {
-      return this.$store.getters.getPlaylist(this.quickBookmarkPlaylistId)
-    },
-    isQuickBookmarkEnabled() {
-      return this.quickBookmarkPlaylist != null
-    },
-    isInQuickBookmarkPlaylist: function () {
-      if (!this.isQuickBookmarkEnabled) { return false }
-
-      return this.quickBookmarkPlaylist.videos.some((video) => {
-        return video.videoId === this.id
-      })
-    },
-    quickBookmarkIconText: function () {
-      if (!this.isQuickBookmarkEnabled) { return this.$t('User Playlists.Add to Playlist') }
-
-      const translationProperties = {
-        playlistName: this.quickBookmarkPlaylist.playlistName,
-      }
-      return this.isInQuickBookmarkPlaylist
-        ? this.$t('User Playlists.Remove from Favorites', translationProperties)
-        : this.$t('User Playlists.Add to Favorites', translationProperties)
-    },
-    quickBookmarkIconTheme: function () {
-      return this.isInQuickBookmarkPlaylist ? 'base favorite' : 'base'
-    },
-    selectQuickBookmarkTargetPromptShown() {
-      return this.$store.getters.getShowSelectQuickBookmarkTargetPrompt
-    },
-  },
-  watch: {
-    selectQuickBookmarkTargetPromptShown(value) {
-      if (value) { return }
-      // Execute on prompt close
-
-      if (this.selectQuickBookmarkTargetPromptCloseCallback == null) { return }
-      this.selectQuickBookmarkTargetPromptCloseCallback()
-    },
   },
   mounted: function () {
     if ('mediaSession' in navigator) {
@@ -359,67 +314,6 @@ export default defineComponent({
       this.showAddToPlaylistPromptForManyVideos({ videos: [videoData] })
     },
 
-    toggleQuickBookmarked() {
-      if (!this.isQuickBookmarkEnabled) {
-        showToast(this.$t('User Playlists["Quick Bookmark Disabled. Pick a Playlist as Quick Bookmark Target"]'))
-        this.selectQuickBookmarkTargetPromptCloseCallback = () => {
-          // Run once only
-          this.selectQuickBookmarkTargetPromptCloseCallback = null
-
-          // Auto add this video to quick bookmark target if target set in prompt
-          if (!this.isQuickBookmarkEnabled) { return }
-
-          // Users don't know the video is in target playlist or not
-          // Assuming they want to add the video
-          // Add it only if not already present in target playlist
-          if (!this.isInQuickBookmarkPlaylist) {
-            this.addToQuickBookmarkPlaylist()
-          }
-        }
-        this.showSelectQuickBookmarkTargetPrompt()
-        return
-      }
-
-      if (this.isInQuickBookmarkPlaylist) {
-        this.removeFromQuickBookmarkPlaylist()
-      } else {
-        this.addToQuickBookmarkPlaylist()
-      }
-    },
-    addToQuickBookmarkPlaylist() {
-      const videoData = {
-        videoId: this.id,
-        title: this.title,
-        author: this.channelName,
-        authorId: this.channelId,
-        description: this.description,
-        viewCount: this.viewCount,
-        lengthSeconds: this.lengthSeconds,
-      }
-
-      this.addVideos({
-        _id: this.quickBookmarkPlaylist._id,
-        videos: [videoData],
-      })
-      // Update playlist's `lastUpdatedAt`
-      this.updatePlaylist({ _id: this.quickBookmarkPlaylist._id })
-
-      // TODO: Maybe show playlist name
-      showToast(this.$t('Video.Video has been saved'))
-    },
-    removeFromQuickBookmarkPlaylist() {
-      this.removeVideo({
-        _id: this.quickBookmarkPlaylist._id,
-        // Remove all playlist items with same videoId
-        videoId: this.id,
-      })
-      // Update playlist's `lastUpdatedAt`
-      this.updatePlaylist({ _id: this.quickBookmarkPlaylist._id })
-
-      // TODO: Maybe show playlist name
-      showToast(this.$t('Video.Video has been removed from your saved list'))
-    },
-
     changeFormat: function(value) {
       this.$emit('change-format', value)
     },
@@ -428,10 +322,6 @@ export default defineComponent({
       'openInExternalPlayer',
       'downloadMedia',
       'showAddToPlaylistPromptForManyVideos',
-      'showSelectQuickBookmarkTargetPrompt',
-      'addVideos',
-      'updatePlaylist',
-      'removeVideo',
     ])
   }
 })
